+++ ios/Classes/InAppWebView/FlutterWebViewController.swift
@@ -58,7 +58,6 @@ public class FlutterWebViewController: NSObject, FlutterPlatformView {
                                    contextMenu: contextMenu,
                                    channel: channel!,
                                    userScripts: userScripts)
-            webView?.disableInputAccessoryView = options.disableInputAccessoryView
         }
         
         methodCallDelegate = InAppWebViewMethodHandler(webView: webView!)